fix: display newStand UI
- Subscribe every components
