+++ src/components/main/Main.ts
@@ -1,3 +1,4 @@
+import { subscribe } from '@/dispatcher';
 import { createElement } from '@utils/domUtils';
 import MainHeader from '@components/main/MainHeader';
 import ListView from '@components/main/listView/ListView';
@@ -8,10 +9,7 @@ import style from '@components/main/Main.module.css';
 type MainProps = {
   gridViewInfo: GridViewInfo;
   subscriptionInfo: string[];
-  mainViewerInfo: {
-    targetMedia: 'total' | 'subscribed';
-    viewer: 'listView' | 'gridView';
-  };
+  mainViewerInfo: MainViewerInfo;
   news: NewsData | null;
   fields: FieldData[];
   listIndex: number;
@@ -46,6 +44,8 @@ export default class Main {
       this.rightButton.element
     );
     this.viewer = props.mainViewerInfo.viewer;
+
+    subscribe(this.updateView.bind(this));
   }
 
   updateView(props: MainProps) {
@@ -55,9 +55,5 @@ export default class Main {
 
       this.element.replaceChild(this.content.element, this.element.childNodes[1]!);
     }
-    this.header.updateView(props);
-    this.content.updateView(props);
-    this.leftButton.updateView(props);
-    this.rightButton.updateView(props);
   }
 }
